Prevent deletion of custom cards during mapping

+++ binkapp/Utils/Protocols/CoreDataRepositoryProtocol.swift
@@ -43,7 +43,15 @@ extension CoreDataRepositoryProtocol {
                     
                     managedObjectIds.forEach { id in
                         if let object = backgroundContext.fetchWithID(type, id: id) {
-                            backgroundContext.delete(object)
+                            
+                            //TODO: Prevent deletion of custom card here
+                            if let membershipCard = object as? CD_MembershipCard {
+                                if membershipCard.membershipPlan?.id != "9999" {
+                                    backgroundContext.delete(object)
+                                }
+                            } else {
+                                backgroundContext.delete(object)
+                            }
                         }
                     }
                 }
